# app/api/v1/gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Count Up","description":"A production-ready Go service deployed on Kubernetes","version":"1.0.0"},"servers":[{"url":"http://localhost:8080"},{"url":"http://localhost:80"}],"paths":{"/":{"get":{"tags":["web"],"summary":"Index web","operationId":"web#Index","responses":{"200":{"description":"OK response.","content":{"text/html":{"schema":{"type":"string","example":"VXQgcGVyc3BpY2lhdGlzIHV0IGVzdCBxdW9kIG1hZ25pLg==","format":"binary"},"example":"QXV0IGxpYmVybyBzb2x1dGEgZG9sb3JlLg=="}}},"401":{"description":"unauthorized: Unauthorized response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/another":{"get":{"tags":["web"],"summary":"Another web","operationId":"web#Another","responses":{"200":{"description":"OK response.","content":{"text/html":{"schema":{"type":"string","example":"RXN0IGVhIG1vbGxpdGlhIHNpbWlsaXF1ZS4=","format":"binary"},"example":"T21uaXMgcGVyZmVyZW5kaXMgaXBzYSBkb2xvciBldW0gZXhlcmNpdGF0aW9uZW0u"}}},"401":{"description":"unauthorized: Unauthorized response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/v1/auth/token":{"post":{"tags":["api"],"summary":"AuthToken api","operationId":"api#AuthToken","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokenRequestBody"},"example":{"access_token":"Omnis cumque est asperiores dolorem.","provider":"google"}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokenResponseBody"},"example":{"token":"Tempore laboriosam maiores autem."}}}},"401":{"description":"unauthorized: Unauthorized response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}},"429":{"description":"existing_increment_request: Too Many Requests response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/v1/counter":{"get":{"tags":["api"],"summary":"CounterGet api","operationId":"api#CounterGet","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CounterInfo"},"example":{"count":1393021988,"last_increment_at":"Quia tenetur illum vitae vitae.","last_increment_by":"Culpa vel enim ut autem.","next_finalize_at":"Illo sapiente."}}}},"401":{"description":"unauthorized: Unauthorized response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}},"429":{"description":"existing_increment_request: Too Many Requests response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}},"security":[{"jwt_header_Authorization":["api"]}]},"post":{"tags":["api"],"summary":"CounterIncrement api","operationId":"api#CounterIncrement","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CounterIncrementRequestBody"},"example":{"user":"Quae cupiditate."}}}},"responses":{"202":{"description":"Accepted response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CounterInfo"},"example":{"count":198148861,"last_increment_at":"Aspernatur vero.","last_increment_by":"Non est qui velit corrupti aliquid placeat.","next_finalize_at":"Atque eligendi quisquam adipisci."}}}},"401":{"description":"unauthorized: Unauthorized response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}},"429":{"description":"existing_increment_request: Too Many Requests response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}},"security":[{"jwt_header_Authorization":["api"]}]}},"/api/v1/openapi.json":{"get":{"tags":["api"],"summary":"Download gen/http/openapi3.json","operationId":"api#/api/v1/openapi.json","responses":{"200":{"description":"File downloaded"}}}},"/echo":{"post":{"tags":["teapot"],"summary":"Echo teapot","operationId":"teapot#Echo","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EchoRequestBody"},"example":{"text":"Tempora repellendus."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EchoRequestBody"},"example":{"text":"Tempore asperiores."}}}}}}},"/login/google":{"get":{"tags":["web"],"summary":"LoginGoogle web","operationId":"web#LoginGoogle","responses":{"302":{"description":"Found response.","headers":{"Location":{"schema":{"type":"string","example":"Quo excepturi eos."},"example":"Quae explicabo."},"Set-Cookie":{"schema":{"type":"string","example":"Labore quidem."},"example":"Officiis consequuntur mollitia provident recusandae."}}},"401":{"description":"unauthorized: Unauthorized response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/login/google/callback":{"get":{"tags":["web"],"summary":"LoginGoogleCallback web","operationId":"web#LoginGoogleCallback","parameters":[{"name":"code","in":"query","allowEmptyValue":true,"required":true,"schema":{"type":"string","example":"Distinctio libero veritatis sint accusamus nulla."},"example":"Expedita sint inventore possimus cumque magni."},{"name":"state","in":"query","allowEmptyValue":true,"required":true,"schema":{"type":"string","example":"Deleniti facere eius eos."},"example":"Facere tempora amet."},{"name":"countup.session","in":"cookie","allowEmptyValue":true,"required":true,"schema":{"type":"string","example":"Omnis distinctio adipisci."},"example":"Aut magni consequatur omnis."}],"responses":{"302":{"description":"Found response.","headers":{"Location":{"schema":{"type":"string","example":"Soluta necessitatibus ad modi esse."},"example":"Expedita commodi quia eos at harum."},"Set-Cookie":{"schema":{"type":"string","example":"Quo aut."},"example":"Beatae voluptatibus ut qui facilis."}}},"401":{"description":"unauthorized: Unauthorized response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/logout":{"get":{"tags":["web"],"summary":"Logout web","operationId":"web#Logout","parameters":[{"name":"countup.session","in":"cookie","allowEmptyValue":true,"required":true,"schema":{"type":"string","example":"Ea voluptas sit consectetur alias libero saepe."},"example":"Ipsam autem."}],"responses":{"302":{"description":"Found response.","headers":{"Location":{"schema":{"type":"string","example":"Error earum ex."},"example":"Et laboriosam odit enim id quod qui."},"Set-Cookie":{"schema":{"type":"string","example":"Molestiae asperiores omnis laborum ea voluptatum dignissimos."},"example":"Voluptas at rerum et alias."}}},"401":{"description":"unauthorized: Unauthorized response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/openapi.json":{"get":{"tags":["teapot"],"summary":"Download gen/http/openapi3.json","operationId":"teapot#/openapi.json","responses":{"200":{"description":"File downloaded"}}}},"/session/token":{"get":{"tags":["web"],"summary":"SessionToken web","operationId":"web#SessionToken","parameters":[{"name":"countup.session","in":"cookie","allowEmptyValue":true,"required":true,"schema":{"type":"string","example":"Ad ullam et."},"example":"Animi in ipsum fuga quod hic."}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokenResponseBody"},"example":{"token":"Itaque laborum et."}}}},"401":{"description":"unauthorized: Unauthorized response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/static/*":{"get":{"tags":["web"],"summary":"Download static/","operationId":"web#/static/*","responses":{"200":{"description":"File downloaded"}}}}},"components":{"schemas":{"AuthTokenRequestBody":{"type":"object","properties":{"access_token":{"type":"string","example":"Eius quasi fugiat quia facilis sint eligendi."},"provider":{"type":"string","example":"google","enum":["google"]}},"example":{"access_token":"Impedit sed.","provider":"google"},"required":["provider","access_token"]},"AuthTokenResponseBody":{"type":"object","properties":{"token":{"type":"string","example":"Corrupti non expedita aliquam illo."}},"example":{"token":"Pariatur est assumenda enim ea."},"required":["token"]},"CounterIncrementRequestBody":{"type":"object","properties":{"user":{"type":"string","example":"Dolorem est doloribus voluptatem hic."}},"example":{"user":"Repudiandae sunt enim alias sunt est dolores."},"required":["user"]},"CounterInfo":{"type":"object","properties":{"count":{"type":"integer","example":1713377072,"format":"int32"},"last_increment_at":{"type":"string","example":"Distinctio rem ipsam rem rerum sunt ex."},"last_increment_by":{"type":"string","example":"Rerum aut adipisci earum sapiente iusto perspiciatis."},"next_finalize_at":{"type":"string","example":"Et rerum pariatur veniam."}},"example":{"count":1838279208,"last_increment_at":"Itaque temporibus eius voluptatem aliquam aut mollitia.","last_increment_by":"Ipsum et sed velit.","next_finalize_at":"Dolores unde eligendi et sint velit."},"required":["count","last_increment_by","last_increment_at","next_finalize_at"]},"EchoRequestBody":{"type":"object","properties":{"text":{"type":"string","example":"Eos voluptatem."}},"example":{"text":"Officiis ut repellendus quo error minima optio."},"required":["text"]},"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"example":{"fault":false,"id":"123abc","message":"parameter 'p' must be an integer","name":"bad_request","temporary":false,"timeout":false},"required":["name","id","message","temporary","timeout","fault"]}},"securitySchemes":{"jwt_header_Authorization":{"type":"http","scheme":"bearer"}}},"tags":[{"name":"api"},{"name":"teapot"},{"name":"web"}]}